requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-Mail
pyodbc
python-dotenv
Werkzeug
openpyxl
email-validator